refactor: `core_adapter` モジュールの切り出し
