+++ voicevox_engine/tts_pipeline/__init__.py
@@ -1,11 +1,10 @@
 from ..core_wrapper import CoreWrapper, load_runtime_lib
 from .make_tts_engines import make_synthesis_engines
-from .tts_engine import CoreAdapter, TTSEngine
+from .tts_engine import TTSEngine
 from .tts_engine_base import TTSEngineBase
 
 __all__ = [
     "CoreWrapper",
-    "CoreAdapter",
     "load_runtime_lib",
     "make_synthesis_engines",
     "TTSEngine",
